refactor: Rename package names from singular to plural form
