+++ src/main/kotlin/com/github/ilovegamecoding/intellijcodexp/listeners/CodeXPEventListener.kt
@@ -1,4 +1,4 @@
-package com.github.ilovegamecoding.intellijcodexp.listener
+package com.github.ilovegamecoding.intellijcodexp.listeners
 
 import com.github.ilovegamecoding.intellijcodexp.enums.Event
 import com.github.ilovegamecoding.intellijcodexp.services.CodeXPService
